pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lvp"
version = "0.1.0"
description = "A universal standard for bandwidth-efficient video upload to multimodal LLMs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "LVP Research Partnership"}
]
keywords = [
    "video",
    "llm",
    "multimodal",
    "preprocessing",
    "bandwidth",
    "ai",
    "machine-learning",
    "claude",
    "gpt-4v",
    "gemini"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
whisper = ["openai-whisper"]
claude = ["anthropic"]
openai = ["openai"]
gemini = ["google-generativeai"]
all = [
    "openai-whisper",
    "anthropic",
    "openai",
    "google-generativeai",
]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
    "ruff",
]

[project.scripts]
lvp = "lvp.cli:main"

[project.urls]
Homepage = "https://github.com/lvp-research/lvp"
Documentation = "https://github.com/lvp-research/lvp#readme"
Repository = "https://github.com/lvp-research/lvp"
Issues = "https://github.com/lvp-research/lvp/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
